requirements.txt
torch
torchvision
timm
gradio
Pillow